+++ tests/src/common.cpp
@@ -1,5 +1,9 @@
 #include "common.hpp"
 
+#include <rapidjson/prettywriter.h>
+#include <rapidjson/stringbuffer.h>
+#include <rapidjson/writer.h>
+
 #include <fstream>
 #include <iostream>
 
@@ -55,4 +59,25 @@ RemoveFile(const std::string &path)
     }
 
     return res;
+}
+
+void
+RJPrettyPrint(const std::shared_ptr<pajlada::Settings::SettingManager> &sm,
+              const std::string &prefix)
+{
+    rapidjson::StringBuffer buffer;
+    rapidjson::PrettyWriter<rapidjson::StringBuffer> writer(buffer);
+    sm->document.Accept(writer);
+
+    std::cout << prefix << buffer.GetString() << '\n';
+}
+
+std::string
+RJStringify(const rapidjson::Value &v)
+{
+    rapidjson::StringBuffer buffer;
+    rapidjson::Writer<rapidjson::StringBuffer> writer(buffer);
+    v.Accept(writer);
+
+    return {buffer.GetString()};
 }